+++ src/packages/excalidraw/CHANGELOG.md
@@ -13,6 +13,12 @@ Please add the latest change on the top under the correct section.
 
 ## Unreleased
 
+- Replace `getElementMap` with [`arrayToMap`](https://github.com/excalidraw/excalidraw/blob/master/src/packages/excalidraw/README.md#arrayToMap).
+
+  #### BREAKING CHANGE
+
+  `getElementMap` util method is now removed.
+
 - Image support.
 
   NOTE: the unreleased API is highly unstable and may change significantly before the next stable release. As such it's largely undocumented at this point. You are encouraged to read through the [PR](https://github.com/excalidraw/excalidraw/pull/4011) description if you want to know more about the internals.